Added integrators that calculate the trajectory of a planet
Used Velocity Verlet to find the evolution of the position and velocity of a planet. The orbits do not look real and work needs to be done on improving this.

Using numpy, the components of many variables e.g. vx and vy have been combined into a single vector. This should improve simulation times.